Fix former upload of 2.0.0-1
The version 2.0.0-1 was used twice and due to that the fixup for
the WAL level is not yet present.

Furthermore the last upload (unmentioned and by mistake) had
needs-root dropped in d/t/control. Tests show that this is needed
so mention that this new upload from git will bring it back.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>